+++ alicloud/resource_alicloud_pvtz_zone_test.go
@@ -10,6 +10,11 @@ import (
 )
 
 func TestAccAlicloudPvtzZone_Basic(t *testing.T) {
+	if !isRegionSupports(PrivateZone) {
+		logTestSkippedBecauseOfUnsupportedRegionalFeatures(t.Name(), PrivateZone)
+		return
+	}
+
 	var zone pvtz.DescribeZoneInfoResponse
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
@@ -33,6 +38,11 @@ func TestAccAlicloudPvtzZone_Basic(t *testing.T) {
 }
 
 func TestAccAlicloudPvtzZone_update(t *testing.T) {
+	if !isRegionSupports(PrivateZone) {
+		logTestSkippedBecauseOfUnsupportedRegionalFeatures(t.Name(), PrivateZone)
+		return
+	}
+
 	var zone pvtz.DescribeZoneInfoResponse
 
 	resource.Test(t, resource.TestCase{
@@ -64,6 +74,11 @@ func TestAccAlicloudPvtzZone_update(t *testing.T) {
 }
 
 func TestAccAlicloudPvtzZone_multi(t *testing.T) {
+	if !isRegionSupports(PrivateZone) {
+		logTestSkippedBecauseOfUnsupportedRegionalFeatures(t.Name(), PrivateZone)
+		return
+	}
+
 	var zone pvtz.DescribeZoneInfoResponse
 
 	resource.Test(t, resource.TestCase{