+++ CHANGELOG.md
@@ -1,12 +1,25 @@
-## 1.14.0 (Unreleased)
+## 1.14.1 (Unreleased)
+## 1.14.0 (August 31, 2018)
 
 FEATURES:
 
+- **New Resource:** `alicloud_vpn_connection` ([#304](https://github.com/terraform-providers/terraform-provider-alicloud/pull/304))
 - **New Resource:** `alicloud_vpn_customer_gateway` ([#299](https://github.com/terraform-providers/terraform-provider-alicloud/pull/299))
 
 IMPROVEMENTS:
 
+- Add 'force' to make key pair affect immediately ([#310](https://github.com/terraform-providers/terraform-provider-alicloud/pull/310))
+- Improve http proxy support ([#307](https://github.com/terraform-providers/terraform-provider-alicloud/pull/307))
+- Add flags to skip tests that use features not supported in all regions ([#306](https://github.com/terraform-providers/terraform-provider-alicloud/pull/306))
+- Improve data source dns_domains test case ([#305](https://github.com/terraform-providers/terraform-provider-alicloud/pull/305))
+- Change SDK config timeout ([#302](https://github.com/terraform-providers/terraform-provider-alicloud/pull/302))
+- Support ClientToken for some request ([#301](https://github.com/terraform-providers/terraform-provider-alicloud/pull/301))
 - Enlarge sdk default timeout to fix some timeout scenario ([#300](https://github.com/terraform-providers/terraform-provider-alicloud/pull/300))
+
+BUG FIXES:
+
+- Fix container cluster SDK timezone error ([#308](https://github.com/terraform-providers/terraform-provider-alicloud/pull/308))
+- Fix network products throttling error ([#303](https://github.com/terraform-providers/terraform-provider-alicloud/pull/303))
 
 ## 1.13.0 (August 28, 2018)